Merge in PE_RBVD_APP-ID-105529_DSG/dto_royal_insurance from feature/RoyalDtosErrorRoyalChangeMap to develop

* commit '6c4fa3b939da2e6c6db8fcecbcf80b606212ea9d':
  PERAR-25146 PERAR-25149 refactor: change data type of httpCode request and response

+++ src/main/java/com/bbva/rbvd/dto/insuranceroyal/error/ErrorRequestDTO.java
@@ -8,7 +8,7 @@ public class ErrorRequestDTO implements Serializable {
     private List<DetailsErrorDTO> details;
     private String typeErrorScope;
     private String channel;
-    private int httpCode;
+    private Long httpCode;
     private String reference;
 
     public List<DetailsErrorDTO> getDetails() {
@@ -35,11 +35,11 @@ public class ErrorRequestDTO implements Serializable {
         this.channel = channel;
     }
 
-    public int getHttpCode() {
+    public Long getHttpCode() {
         return httpCode;
     }
 
-    public void setHttpCode(int httpCode) {
+    public void setHttpCode(Long httpCode) {
         this.httpCode = httpCode;
     }
